add ProductsSolutionsLanding miniapp component
Note that this replaces /apps/bmc/components/content/related-CTAs/related-CTAs-landing
with /apps/bmc/components/content/products-solutions-landing/topic/. The former was
planned for but made no sense in practice.

Also, this miniapp component is not 100%, still needs some code to pull out the all-products lists
which live at the bottom of the markup.